ClusterMonitor: split to its own file
Moved ClusterMonitor ands its dependencies as-is to a new file.
